HUDI-494 fix incorrect record size estimation
